Move alternate builder to KeyPool.Builder, deprecate KeyPoolBuilder
