+++ core/src/main/scala/org/typelevel/keypool/KeyPool.scala
@@ -191,7 +191,7 @@ object KeyPool {
   }
 
   private[keypool] def state[F[_]: Functor, A, B](
-      kpVar: Ref[F, PoolMap[A, B]]
+      kpVar: Ref[F, PoolMap[A, (B, F[Unit])]]
   ): F[(Int, Map[A, Int])] =
     kpVar.get.map(pm =>
       pm match {
@@ -243,7 +243,7 @@ object KeyPool {
                 val (l_, mx) = addToList(now, kp.kpMaxPerKey(k), (r, destroy), l)
                 val cnt_ = idleCount + mx.fold(1)(_ => 0)
                 val m_ = PoolMap.open(cnt_, m + (k -> l_))
-                (m_, mx.fold(Applicative[F].unit)(r => destroy))
+                (m_, mx.fold(Applicative[F].unit)(_ => destroy))
             }
       }
 
@@ -283,5 +283,114 @@ object KeyPool {
           } yield out
       }
     } yield new Managed(resource._1, optR.isDefined, releasedState)
+  }
+
+  final class Builder[F[_]: Temporal, A, B] private (
+      val kpRes: A => Resource[F, B],
+      val kpDefaultReuseState: Reusable,
+      val idleTimeAllowedInPool: Duration,
+      val kpMaxPerKey: A => Int,
+      val kpMaxTotal: Int,
+      val onReaperException: Throwable => F[Unit]
+  ) {
+    private def copy(
+        kpRes: A => Resource[F, B] = this.kpRes,
+        kpDefaultReuseState: Reusable = this.kpDefaultReuseState,
+        idleTimeAllowedInPool: Duration = this.idleTimeAllowedInPool,
+        kpMaxPerKey: A => Int = this.kpMaxPerKey,
+        kpMaxTotal: Int = this.kpMaxTotal,
+        onReaperException: Throwable => F[Unit] = this.onReaperException
+    ): Builder[F, A, B] = new Builder[F, A, B](
+      kpRes,
+      kpDefaultReuseState,
+      idleTimeAllowedInPool,
+      kpMaxPerKey,
+      kpMaxTotal,
+      onReaperException
+    )
+
+    def doOnCreate(f: B => F[Unit]): Builder[F, A, B] =
+      copy(kpRes = { (k: A) => this.kpRes(k).flatMap(v => Resource.eval(f(v).attempt.void.as(v))) })
+
+    def doOnDestroy(f: B => F[Unit]): Builder[F, A, B] =
+      copy(kpRes = { (k: A) =>
+        this.kpRes(k).flatMap(v => Resource.make(Applicative[F].unit)(_ => f(v).attempt.void).as(v))
+      })
+
+    def withDefaultReuseState(defaultReuseState: Reusable) =
+      copy(kpDefaultReuseState = defaultReuseState)
+
+    def withIdleTimeAllowedInPool(duration: Duration) =
+      copy(idleTimeAllowedInPool = duration)
+
+    def withMaxPerKey(f: A => Int): Builder[F, A, B] =
+      copy(kpMaxPerKey = f)
+
+    def withMaxTotal(total: Int): Builder[F, A, B] =
+      copy(kpMaxTotal = total)
+
+    def withOnReaperException(f: Throwable => F[Unit]) =
+      copy(onReaperException = f)
+
+    def build: Resource[F, KeyPool[F, A, B]] = {
+      def keepRunning[Z](fa: F[Z]): F[Z] =
+        fa.onError { case e => onReaperException(e) }.attempt >> keepRunning(fa)
+      for {
+        kpVar <- Resource.make(
+          Ref[F].of[PoolMap[A, (B, F[Unit])]](PoolMap.open(0, Map.empty[A, PoolList[(B, F[Unit])]]))
+        )(kpVar => KeyPool.destroy(kpVar))
+        _ <- idleTimeAllowedInPool match {
+          case fd: FiniteDuration =>
+            val nanos = 0.seconds.max(fd)
+            Resource.make(
+              Concurrent[F].start(keepRunning(KeyPool.reap(nanos, kpVar, onReaperException)))
+            )(_.cancel)
+          case _ =>
+            Applicative[Resource[F, *]].unit
+        }
+      } yield new KeyPool.KeyPoolConcrete(
+        kpRes,
+        kpDefaultReuseState,
+        kpMaxPerKey,
+        kpMaxTotal,
+        kpVar
+      )
+    }
+
+  }
+
+  object Builder {
+    def apply[F[_]: Temporal, A, B](
+        res: A => Resource[F, B]
+    ): Builder[F, A, B] = new Builder[F, A, B](
+      res,
+      Defaults.defaultReuseState,
+      Defaults.idleTimeAllowedInPool,
+      Defaults.maxPerKey,
+      Defaults.maxTotal,
+      Defaults.onReaperException[F]
+    )
+
+    def apply[F[_]: Temporal, A, B](
+        create: A => F[B],
+        destroy: B => F[Unit]
+    ): Builder[F, A, B] = new Builder[F, A, B](
+      a => Resource.make(create(a))(destroy),
+      Defaults.defaultReuseState,
+      Defaults.idleTimeAllowedInPool,
+      Defaults.maxPerKey,
+      Defaults.maxTotal,
+      Defaults.onReaperException[F]
+    )
+
+    private object Defaults {
+      val defaultReuseState = Reusable.Reuse
+      val idleTimeAllowedInPool = 30.seconds
+      def maxPerKey[K](k: K): Int = Function.const(100)(k)
+      val maxTotal = 100
+      def onReaperException[F[_]: Applicative] = { (t: Throwable) =>
+        Function.const(Applicative[F].unit)(t)
+      }
+    }
   }
 }
